Package.json name of the sdk changed
[![Build Status](https://travis-ci.org/joemccann/dillinger.svg?branch=master)](https://travis-ci.org/joemccann/dillinger)

`prodio-notification-sdk` is a service module to send emails, sms, and push app notifications to both mobile phones and web browsers. This service is easy to integrate with your existing product or infrastructure as it comes with its on environment and dependencies.

This project is started with an aim to reduce implementing and re-architecting common product features/requirements like notification service & thus increasing productivity and more focus on business goals. Goal of this service is to bundle and provide common features related to notification services like sending emails, sms, notifications to users, notification logs and history, events and templates creation for easier integration with web and mobile Apps.

1. Create User

$ npm install prodio-notification-sdk --save

Sdk module provides a easy and fast way to build notifications in the product.Examples like New Member added to a team, Someone messaged you, one of your friends liked your post, etc are events. Keeping event-driven architecture in mind this service works or reacts to events, which integration with others services easier. You can trigger emails, or send notifications in response to an event.

This application will run as a separate micro-service independent of your product services which makes it easier to implement, debug and test.

```JSX

	var  notificationModule = require('prodio-notification-sdk');
	//config
	const config = {
	 "BASE_URL":"API GATE_WAY LINK"
	};
	var  notificationObject = notificationModule(config);
	//create user in notification module
	const meta_info = {
	     "user_name":"user_name",
		 "email":"user_email",
	    }
	const  payload = {
	"userId":"user_id",
	"meta_info":meta_info,
	"event_name":"CREATE_USER"
	};

let  createUser = notificationModule.createUser(payload);

```

| Prop | Type | Description|

|:--------------------:|:----------------------------------------------------------------------------:|:-----------|

| `userId` | string |unique Identity Id of the user created| |

| `Email` | string | Email address of the user created. |

| `Name` | string| Name of the user created |
| Event_Name | string | Event name for creating events for  |

method will create the user in the product which will we able to receive notifications , receive emails and messages.